+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/bridgeless/BridgelessReactPackage.java
@@ -5,9 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package com.facebook.react;
+package com.facebook.react.bridgeless;
 
 import com.facebook.infer.annotation.Nullsafe;
+import com.facebook.react.TurboReactPackage;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.devsupport.LogBoxModule;
